% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{adapt_test_options}
\alias{adapt_test_options}
\title{Adaptive test options}
\usage{
adapt_test_options(
  next_item.criterion = "MFI",
  next_item.estimator = "BM",
  next_item.prior_dist = "norm",
  next_item.prior_par = c(0, 1),
  final_ability.estimator = "BM",
  constrain_answers = FALSE,
  avoid_duplicates = NULL,
  cb_control = NULL,
  cb_group = NULL,
  eligible_first_items = NULL,
  notify_duration = 5,
  mixed_effects_model = NULL,
  continuous_response = FALSE,
  dv_name = " ",
  fixed_effects = c("fixed_effect1", "fixed_effect2")
)
}
\arguments{
\item{next_item.criterion}{(Character scalar, default = "MFI")
Criterion by which the next item is selected;
see the \code{criterion} argument of \code{\link[catR]{nextItem}}.}

\item{next_item.estimator}{(Character scalar, default = "BM")
Ability scoring method used for item selection;
see the \code{method} argument of \code{\link[catR]{nextItem}}.
Only \code{"BM"}, \code{"ML"}, \code{"WL"}, and \code{"EAP"}
are supported.}

\item{next_item.prior_dist}{(Character scalar, default = "norm")
Class of prior distribution used when estimating abilities
for item selection;
see the \code{priorDist} argument of \code{\link[catR]{nextItem}}.}

\item{next_item.prior_par}{(Numeric vector, length 2, default = c(0, 1))
Prior distribution used when estimating abilities for item selection;
see the \code{priorPar} argument of \code{\link[catR]{nextItem}}.}

\item{final_ability.estimator}{(Character scalar, default = "BM")
Ability scoring method used for final ability estimation;
see the \code{method} argument of \code{\link[catR]{nextItem}}.
The prior distribution (if relevant) remains the same as that used
for item selection, i.e. that specified by
\code{next_item.prior_dist} and \code{next_item.prior_par}.
Only \code{"BM"}, \code{"ML"}, \code{"WL"}, and \code{"EAP"}
are supported.}

\item{constrain_answers}{(Boolean; default = FALSE)
If \code{TRUE}, item selection actively balances
the available response options.
This only makes sense when the item bank contains only a small
number of unique answers.}

\item{avoid_duplicates}{(\code{NULL} or character scalar)
If not \code{NULL}, should specify a column of the item bank
defining a categorical item feature for which duplicates
should be avoided during item selection.}

\item{cb_control}{Content-balancing argument passed to \code{\link[catR]{nextItem}}.}

\item{cb_group}{Content-balancing argument passed to \code{\link[catR]{nextItem}}.}

\item{eligible_first_items}{(\code{NULL} or integerish vector)
If not \code{NULL},
lists the eligible items for the first item in the test,
where each item is identified by its 1-indexed row number
in \code{item_bank} (see \code{\link{adapt_test}}).
For example, \code{c(2, 3, 4)} means that the
first item will be drawn from rows 2, 3, 4 of the item bank).}

\item{notify_duration}{(Numeric scalar)
Defines the length of time for which item-by-item feedback messages
persist on screen, in seconds.
This is only relevant when the test is taken in admin or demo mode.}

\item{mixed_effects_model}{A mixed effects model of class lmerModLmerTest.}

\item{continuous_response}{Is response continuous?}

\item{dv_name}{Name of dependent variable.}

\item{fixed_effects}{Character vector of names of fixed effects.}
}
\value{
A list to be passed to the \code{opt} argument
of \code{\link{adapt_test}}.
}
\description{
Creates a list of psychTestRCAT options for use in
the \code{\link{adapt_test}} function.
}
