% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{cat.feedback.irt}
\alias{cat.feedback.irt}
\title{Feedback on IRT score}
\usage{
cat.feedback.irt(text = "feedback_irt",
  dict = psychTestRCAT::ptrcat_dict, next_button = NULL,
  digits_irt_score = 3L, digits_irt_sem = 3L)
}
\arguments{
\item{text}{(Character scalar)
Text to display to the participant. This text will be treated as an
internationalisation key for the dictionary contained in the \code{dict}
argument; if the key is not present in the dictionary, the text
will be displayed as is.}

\item{dict}{Internationalisation dictionary (see \code{\link[psychTestR]{i18n_dict}}).}

\item{next_button}{(NULL or a character scalar or an object of class "shiny.tag")
If NULL, no next button is shown
(typically because the test has completed).
Otherwise, a button to progress to the next page is created,
displaying the content of this argument.}

\item{digits_irt_score}{(Numeric scalar)
Number of digits to which the IRT score should be rounded.}

\item{digits_irt_sem}{(Numeric scalar)
Number of digits to which the IRT standard error should be rounded.}
}
\value{
A timeline segment suitable for inclusion in a psychTestR timeline
directly after \code{\link{adapt_test}}.
}
\description{
Displays the participant's IRT score and associated standard error
to the participant. Most participants won't know what these scores mean;
however, this feedback option can be useful to experimenters
conducting in-lab studies, because it means they can write down
the final scores manually.
}
