% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_mixed_effects.R
\name{compute_new_ranef_full_model}
\alias{compute_new_ranef_full_model}
\title{Compute a participant random effect intercept, based on their item scores}
\usage{
compute_new_ranef_full_model(
  y,
  N,
  step.cont.loc.var,
  tonalness,
  log_freq,
  min,
  max,
  fixed_effect_intercept,
  sigma
)
}
\arguments{
\item{sigma}{}
}
\value{

}
\description{
Compute a participant random effect intercept, based on their item scores
}
