% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythmic_model.R
\name{compute_new_ranef_rhythmic_model}
\alias{compute_new_ranef_rhythmic_model}
\title{Compute a participant random effect intercept, based on the scores they got on different items (with different difficulty features)}
\usage{
compute_new_ranef_rhythmic_model(
  y,
  N,
  step.cont.loc.var,
  log_freq,
  d.entropy,
  i.entropy,
  min,
  max,
  fixed_effect_intercept,
  sigma
)
}
\arguments{
\item{sigma}{}
}
\description{
Compute a participant random effect intercept, based on the scores they got on different items (with different difficulty features)
}
