% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-item-bank.R
\name{demo_item_bank}
\alias{demo_item_bank}
\title{Demo item bank}
\usage{
demo_item_bank(n = 20, questions = 1:4, answers = 1:2)
}
\arguments{
\item{n}{(Integerish scalar) Number of questions to include in the item bank.}

\item{questions}{Vector of questions to sample from.}

\item{answers}{Vector of answers to sample from.}
}
\value{
A generated item bank in the form of a data frame,
suitable for passing to \code{\link{adapt_test}}.
}
\description{
This function creates an example item bank
with randomly generated difficulty and discrimination parameters.
}
