% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-show-item-list.R
\name{demo_show_item_list}
\alias{demo_show_item_list}
\title{Demo: show_item of class list}
\usage{
demo_show_item_list()
}
\description{
This is a demonstration of an adaptive test with a variable number
of pages for each item.
This is achieved by using a \code{show_item} argument corresponding
to a list of test elements.
The crucial trick is using a \code{loop_while} function to
repeat a 'reactive page' until the item is complete,
and a local variable \code{counter} to keep track of the iteration number.
}
