% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{get_allowed_items}
\alias{get_allowed_items}
\title{The item must match correct_answers (if specified),
and it must avoid duplicates in opt$avoid_duplicates.
If this is not possible, the item is chosen randomly, with a warning.}
\usage{
get_allowed_items(test_state, item_bank, opt)
}
\description{
The item must match correct_answers (if specified),
and it must avoid duplicates in opt$avoid_duplicates.
If this is not possible, the item is chosen randomly, with a warning.
}
