% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effects.R
\name{get_current_ability_estimate_mixed_effects}
\alias{get_current_ability_estimate_mixed_effects}
\title{Current ability via mixed effects modelling estimate}
\usage{
get_current_ability_estimate_mixed_effects(
  test_state,
  opt,
  estimator = opt$next_item.estimator,
  mixed_effects_model = opt$mixed_effects_model
)
}
\arguments{
\item{test_state}{A \code{test_state} object as provided to
stopping rules in \code{\link{new_stopping_rule}}.}

\item{opt}{Test options as defined by \code{\link{adapt_test_options}}.}

\item{estimator}{The estimation method to use when computing
the current ability estimate;
see the \code{next_item.estimator} argument in
\code{\link{adapt_test_options}}.}
}
\value{
A numeric scalar corresponding to the current ability estimate,
with its standard error provided as the \code{sem} attribute.
}
\description{
Returns the current ability estimate.
}
