% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{get_item_number}
\alias{get_item_number}
\alias{get_item_number.item}
\title{Get item number}
\usage{
get_item_number(x)

\method{get_item_number}{item}(x)
}
\arguments{
\item{x}{An item object as passed to the \code{show_item} function
(see \code{\link{adapt_test}}).}
}
\value{
(Integer scalar) The item's position in the test; for example,
the second item in the test would yield the integer \code{2L}.
}
\description{
Returns an item's position in the test.
}
