% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{get_num_items_administered}
\alias{get_num_items_administered}
\title{Number of items administered}
\usage{
get_num_items_administered(test_state)
}
\arguments{
\item{test_state}{A \code{test_state} object; see \code{\link{new_stopping_rule}}.}
}
\value{
(Integer scalar)
The number of items administered so far in the test.
}
\description{
Returns the number of items administered so far in the test.
}
