% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{get_num_items_in_test}
\alias{get_num_items_in_test}
\alias{get_num_items_in_test.stopping_rule}
\alias{get_num_items_in_test.item}
\alias{get_num_items_in_test.test_state}
\title{Number of items in test}
\usage{
get_num_items_in_test(x)

\method{get_num_items_in_test}{stopping_rule}(x)

\method{get_num_items_in_test}{item}(x)

\method{get_num_items_in_test}{test_state}(x)
}
\arguments{
\item{x}{Object from which to extract the information.}
}
\value{
(Integer scalar)
The total number of items in the test,
including those that have yet to be administered.
}
\description{
Returns the total number of items in the test,
including those that have yet to be administered.
Methods exist for:
\itemize{
\item \code{stopping_rule} (see \code{\link{new_stopping_rule}}).
\item \code{item} (as passed to \code{show_item} in \code{\link{adapt_test}}).
\item \code{test_state} (as passed to stopping rules,
see \code{\link{new_stopping_rule}}).
}
}
