% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{new_stopping_rule}
\alias{new_stopping_rule}
\title{New stopping rule}
\usage{
new_stopping_rule(f, num_items_in_test = NULL)
}
\arguments{
\item{f}{Function of the form \code{function(test_state) {...}},
which will be called whenever a new item is about to be administered.
\code{test_state} is a data frame representing the current status of
the adaptive test (items administered, participant responses, etc.).
Currently the only method defined for \code{test_state} objects
is \code{\link{get_num_items_administered}};
other extractors need to be defined manually.}

\item{num_items_in_test}{(NULL or integerish scalar)
If the number of items in the test is known,
providing it in this argument allows it to be used by other test features,
e.g. to tell the participant their progress through the test.}
}
\value{
An object of class \code{stopping_rule} which can be
passed to the \code{stopping_rule} argument of
\code{\link{adapt_test}}.
}
\description{
Creates a new stopping rule for the adaptive test.
}
\details{
A stopping rule defines a condition that is checked each time
a new item is about to be administered.
}
