% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrhythmic_model.R
\name{predict_based_on_mixed_effects_arrhythmic_model}
\alias{predict_based_on_mixed_effects_arrhythmic_model}
\title{Predict based on arrhythmic mixed-effects model}
\usage{
predict_based_on_mixed_effects_arrhythmic_model(model, new_data)
}
\arguments{
\item{new_data}{}
}
\value{

}
\description{
Predict based on arrhythmic mixed-effects model
}
