% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic_mixed_effects.R
\name{predict_based_on_mixed_effects_model}
\alias{predict_based_on_mixed_effects_model}
\title{Predict based on a mixed-effects model}
\usage{
predict_based_on_mixed_effects_model(model, new_data)
}
\arguments{
\item{new_data}{}
}
\value{

}
\description{
Predict based on a mixed-effects model
}
