% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhythmic_model.R
\name{predict_based_on_mixed_effects_rhythmic_model}
\alias{predict_based_on_mixed_effects_rhythmic_model}
\title{Predict based on rhythmic mixed-effects model}
\usage{
predict_based_on_mixed_effects_rhythmic_model(model, new_data)
}
\arguments{
\item{new_data}{}
}
\description{
Predict based on rhythmic mixed-effects model
}
