% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychTestRCAT-package.R
\docType{package}
\name{psychTestRCAT-package}
\alias{psychTestRCAT}
\alias{psychTestRCAT-package}
\title{psychTestRCAT}
\description{
This package supports the creation of adaptive tests
within the psychTestR computerised testing framework.
}
\details{
Start creating an adaptive test with the \code{\link{adapt_test}} function.
Also see the following example implementations:
\itemize{
\item \href{http://github.com/pmcharrison/mdt}{mdt} -
Melody discrimination test.
\item \href{http://github.com/pmcharrison/mpt}{mpt} -
Mistuning perception test.
\item \href{http://github.com/pmcharrison/cabat}{cabat} -
Beat perception test.
\item \href{http://github.com/pmcharrison/piat}{piat} -
Pitch imagery test.
}
}
\author{
\strong{Maintainer}: Peter Harrison \email{p.m.c.harrison@qmul.ac.uk}

}
\keyword{internal}
