% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychTestRCAT-package.R
\docType{data}
\name{ptrcat_dict}
\alias{ptrcat_dict}
\title{psychTestRCAT dictionary}
\description{
This object defines a dictionary (see \code{\link[psychTestR]{i18n_dict}})
for internationalising psychTestRCAT.
}
\keyword{data}
