% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{stopping_rule.num_items}
\alias{stopping_rule.num_items}
\title{Stopping rule (num items)}
\usage{
stopping_rule.num_items(n)
}
\arguments{
\item{n}{(Integerish scalar)
Number of items after which the test should terminate.}
}
\value{
An object of class \code{stopping_rule} which can be
passed to the \code{stopping_rule} argument of
\code{\link{adapt_test}}.
}
\description{
Creates a stopping rule where the test finishes once a set number
of items have been administered.
}
